Improved imports and targeting
Use either `import { Bonjour } from 'bonjour-service' or import Bonjour from 'bonjour-service'`
