+++ tsconfig.json
@@ -1,17 +1,20 @@
 {
     "compilerOptions": {
       "target": "es6",
-      "module": "commonjs",
+      "module": "node16",
       "declaration": true,
       "sourceMap": true,
-       "outDir": "./dist",
+      "outDir": "./dist",
       "strict": true,
-      "moduleResolution": "node",
+      "moduleResolution": "node16",
       "baseUrl": ".",
       "allowSyntheticDefaultImports": true,
       "esModuleInterop": true,
       "skipLibCheck": true,
-      "forceConsistentCasingInFileNames": true
+      "forceConsistentCasingInFileNames": true,
+      "types": [
+        "node"
+      ]
     }
   }
